Revert "refactor(stop_filter): prefix package and namespace with autoware (#7789)"
This reverts commit fbbc44ceb33c423c00ba3caa826ff77c8958d920.
