+++ localization/stop_filter/CMakeLists.txt
@@ -1,12 +1,11 @@
 cmake_minimum_required(VERSION 3.14)
-project(autoware_stop_filter)
+project(stop_filter)
 
 find_package(autoware_cmake REQUIRED)
 autoware_package()
 
 ament_auto_add_library(${PROJECT_NAME} SHARED
   src/stop_filter.cpp
-  src/stop_filter.hpp
 )
 
 rclcpp_components_register_node(${PROJECT_NAME}
