+++ localization/stop_filter/launch/stop_filter.launch.xml
@@ -1,9 +1,9 @@
 <launch>
-  <arg name="param_path" default="$(find-pkg-share autoware_stop_filter)/config/stop_filter.param.yaml"/>
+  <arg name="param_path" default="$(find-pkg-share stop_filter)/config/stop_filter.param.yaml"/>
   <arg name="input_odom_name" default="ekf_odom"/>
   <arg name="output_odom_name" default="stop_filter_odom"/>
   <arg name="debug_stop_flag" default="debug/stop_flag"/>
-  <node pkg="autoware_stop_filter" exec="autoware_stop_filter_node" output="both">
+  <node pkg="stop_filter" exec="stop_filter_node" output="both">
     <remap from="input/odom" to="$(var input_odom_name)"/>
 
     <remap from="output/odom" to="$(var output_odom_name)"/>
