+++ localization/stop_filter/schema/stop_filter.schema.json
